--- edu.kit.ipd.sdq.kamp.ruledsl/src/edu/kit/ipd/sdq/kamp/ruledsl/generator/IRule.java
@@ -1,5 +1,0 @@
-package edu.kit.ipd.sdq.kamp.ruledsl.generator;
-
-public interface IRule {
-	void apply();
-}
