+++ edu.kit.ipd.sdq.kamp.ruledsl/src/edu/kit/ipd/sdq/kamp/ruledsl/service/IRule.java
@@ -1,0 +1,5 @@
+package edu.kit.ipd.sdq.kamp.ruledsl.service;
+
+public interface IRule {
+	void apply();
+}
