+++ edu.kit.ipd.sdq.kamp.ruledsl/src/edu/kit/ipd/sdq/kamp/ruledsl/service/IRuleProvider.java
@@ -1,4 +1,4 @@
-package edu.kit.ipd.sdq.kamp.ruledsl.generator;
+package edu.kit.ipd.sdq.kamp.ruledsl.service;
 
 public interface IRuleProvider {
 	void applyAllRules();
